Add imports

+++ src/hotspot/share/runtime/deoptimization.cpp
@@ -2091,7 +2091,7 @@ JRT_ENTRY(void, Deoptimization::uncommon_trap_inner(JavaThread* current, jint tr
     // Load class if necessary
     if (unloaded_class_index >= 0) {
       constantPoolHandle constants(current, trap_method->constants());
-      if (unloaded_class_index < 0) {
+      if ((int) unloaded_class_index < 0) {
         fprintf(stderr, "Why? %d\n", unloaded_class_index);
       }
       load_class_by_index(constants, unloaded_class_index, THREAD);
